+++ lib/shrine/plugins/activerecord.rb
@@ -35,12 +35,14 @@ class Shrine
               attacher.save if attacher.changed?
             end
 
-            model.after_commit on: [:create, :update] do
-              attacher = send("#{name}_attacher")
-              attacher.finalize if attacher.changed?
+            [:create, :update].each do |action|
+              model.after_commit on: action do
+                attacher = send("#{name}_attacher")
+                attacher.finalize if attacher.changed?
+              end
             end
 
-            model.after_commit on: [:destroy] do
+            model.after_commit on: :destroy do
               send("#{name}_attacher").destroy
             end
           end